ui-redesign: Login, fix: CustomButtons
